+++ NEWS.md
@@ -1,3 +1,10 @@
+# poissoned 0.1.2
+
+* Random inititialisation point is now confined to the central area to avoid 
+  some edge cases.  This fixes some tiling artefacts.
+* Fixed up an error in how corner points were being mirrored as boundary points.
+  This fixes some tiling artefacts.
+
 # poissoned 0.1.1
 
 * Added arguments `xinit`, `yinit` to allow the user to specify the initial seed point